+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Echterhof 2026 Pseudomonas bacteriophage whole-body PBPK ([doi:10.1128/aac.01506-25](https://doi.org/10.1128/aac.01506-25)) -- three models: CD-1 mice (the fitted species), plus rat and human interspecies projections.
+- Add Kim 2026 midazolam ([doi:10.1097/ALN.0000000000005811](https://doi.org/10.1097/ALN.0000000000005811)) -- adults on venoarterial ECMO and after decannulation.
 
 - Add Siebinga 2023 [68Ga]Ga- and [177Lu]Lu-HA-DOTATATE ([doi:10.1186/s40658-023-00565-4](https://doi.org/10.1186/s40658-023-00565-4)) - adults with neuroendocrine tumors.